requirements.txt
amqp==2.2.2
billiard==3.5.0.3
boto3==1.5.1
botocore==1.8.15
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0
django-ckeditor==5.4.0
django-decouple==2.1
django-import-export==0.7.0
django-js-asset==1.0.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
docutils==0.14
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
jdcal==1.3
jmespath==0.9.3
kombu==4.1.0
oauthlib==2.0.6
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python-dateutil==2.6.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
sorl-thumbnail==12.4.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
